fixed : most of the compiler messages (on windoes)       : still in : printf format stuff (seems to be platform dependend)                    few copy to temp    (have to have a closer look)

+++ FieldContainer/OSGAttachmentContainerPtr.cpp
@@ -254,7 +254,7 @@ ostream &OSG::operator <<(      ostream                 &os,
     {
         os << hex << "NodePtr 0x"
            << &fc << dec << ":" << fc->getType().getName() << "Ptr(0x"
-           << hex << (int) (&(*fc)) << dec << ")";
+           << hex << (&(*fc)) << dec << ")";
     }
 
     return os;
